+++ firmware/examples/meshuptime/WebTask.h
@@ -349,6 +349,7 @@ private:
   void handleAnnounceSet();
   void handleAnnounceDel();
   void handleAnnounceTest();
+  void handleAnnounceGap();
 
   /* Companions (v2.4.0):
    *  /companions.json (GET) : {companions:[{name,pubkey,lat,lon,seen}...]} -- ook
@@ -428,6 +429,7 @@ private:
   friend void web_route_annset();
   friend void web_route_anndel();
   friend void web_route_anntest();
+  friend void web_route_anngap();
   friend void web_route_companionsjson();
   friend void web_route_companion();
   friend void web_route_messagesjson();